requirements.txt
shazamio==0.4.0.1
python-dotenv==1.0.0
FastAPI==0.109.0
uvicorn
streamlink==6.5.0
fs~=2.4.16
pydantic
requests==2.31.0
boto3